Rename MarkdownParser to MarkdownProcessor::Parse
