+++ app/controllers/articles_controller.rb
@@ -87,7 +87,7 @@ class ArticlesController < ApplicationController
     begin
       fixed_body_markdown = MarkdownFixer.fix_for_preview(params[:article_body])
       parsed = FrontMatterParser::Parser.new(:md).call(fixed_body_markdown)
-      parsed_markdown = MarkdownParser.new(parsed.content, source: Article.new, user: current_user)
+      parsed_markdown = MarkdownProcessor::Parse.new(parsed.content, source: Article.new, user: current_user)
       processed_html = parsed_markdown.finalize
     rescue StandardError => e
       @article = Article.new(body_markdown: params[:article_body])
