+++ app/models/user_subscription.rb
@@ -58,7 +58,7 @@ class UserSubscription < ApplicationRecord
   end
 
   def liquid_tags_used
-    MarkdownParser.new(
+    MarkdownProcessor::Parse.new(
       user_subscription_sourceable.body_markdown,
       source: user_subscription_sourceable,
       user: user_subscription_sourceable.user,
